test(accordion): migrate basic test
